Remove functions that were moved to pygeoops
