+++ geofileops/fileops.py
@@ -22,11 +22,11 @@ from geopandas.io import file as gpd_io_file
 import numpy as np
 from osgeo import gdal
 import pandas as pd
+from pygeoops import GeometryType, PrimitiveType  # noqa: F401
 import pyogrio
 import pyproj
 
-from geofileops.util.geometry_util import GeometryType, PrimitiveType  # noqa: F401
-from geofileops.util import geoseries_util
+from geofileops.util import _geoseries_util
 from geofileops.util import _io_util
 from geofileops.util import _ogr_util
 from geofileops.util import _ogr_sql_util
@@ -1570,7 +1570,7 @@ def _to_file_fiona(
             # doesn't like > 1 type in a layer
             if schema is None or (len(gdf) > 0 and schema["geometry"] != "None"):
                 gdf = gdf.copy()
-                gdf.geometry = geoseries_util.harmonize_geometrytypes(
+                gdf.geometry = _geoseries_util.harmonize_geometrytypes(
                     gdf.geometry, force_multitype=force_multitype
                 )
                 assert isinstance(gdf, gpd.GeoDataFrame)
